refactor(compute): rename `compute logs` to `compute events`
The OSS endpoint this command calls returns Fly machine lifecycle
events (start/stop/exit/restart) — not container stdout/stderr.
Calling it `compute logs` was misleading; `compute events` is honest.

When real container stdout/stderr lands later, the freshly-vacated
`compute logs` command name will be used for actual log streaming.

Mild breaking change — `npx`-based usage cache means most users won't
notice on the next invocation. Tracks the parallel rename in OSS PR
that moves /api/compute/services/:id/logs → /events.
